// src/components/BookWriter.module.css.d.ts
declare const styles: {
  bookWriter: string;
  eventsSection: string;
  noEvents: string;
  eventsList: string;
  eventItem: string;
  observationIndicator: string;
  observedIcon: string;
  unobservedIcon: string;
  eventDescription: string;
  form: string;
  claimsSection: string;
  claimInput: string;
  select: string;
  textarea: string;
  addButton: string;
  removeButton: string;
  submitButton: string;
  errorMessage: string;
};

export = styles;